MENU CRUD

+++ routes/menus.js
@@ -1,26 +1,19 @@
 const express = require('express');
 const router = express.Router();
-// const {MongoClient} = require("mongodb");
 const Menu = require('../models/menu');
-
-// const mongoose = require('mongoose');
-
-
-// // mongoose.connect('mongodb://localhost/Menus', {useNewUrlParser: true, useUnifiedTopology: true });
-
-// const db = mongoose.connection;
-// db.on('error', console.error.bind(console, 'connection error:'));
-// db.once('open', function() {
  
   router.get('/', function(req, res, next) {
+
     Menu.find()
     .then((menus) => {
       res.send(menus);
     })
     .catch(next);
+
   });
 
   router.post('/',  async function(req, res, next) {
+
     Menu.findOne({menuname: req.body.menuname})
     .then((result) => {
       if (!result) {
@@ -38,21 +31,40 @@ const Menu = require('../models/menu');
   });
 
   router.put('/:menuname', function(req, res, next) {
-    Menu.findOneAndUpdate({"menuname": req.params.menuname}, {$push: {"drinksList": req.body}})
-    .then((menu) => {
-      res.send(req.body);
+
+    Menu.findOne({menuname: req.params.menuname})
+    .then((result) => {
+      if (result) {
+        Menu.findOneAndUpdate({"menuname": req.params.menuname}, {$push: {"drinksList": req.body}})
+        .then((menu) => {
+          res.send(req.body);
+        })
+        .catch(next);
+      } else {
+        throw("A menu with this name does not exist");
+      }
     })
     .catch(next);
+    
   });
 
   router.delete('/:menuname', function(req, res, next) {
-    Menu.findOneAndDelete({ menuname: req.params.menuname })
-    .then((menu) => {
-      res.status(200).send(menu);
+    
+    Menu.findOne({menuname: req.params.menuname})
+    .then((result) => {
+      if (result) {
+        Menu.findOneAndDelete({ menuname: req.params.menuname })
+        .then((menu) => {
+          res.status(200).send(menu);
+        })
+        .catch(next);
+      } else {
+        throw("A menu with this name does not exist");
+      }
     })
     .catch(next);
+  
   });
 
-// });
 
 module.exports = router;
